fix : 칼질
