+++ lib/2025-04-04/data_source/mask_store_data_source_impl.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 import 'package:modu_3_dart_study/2025-04-04/dto/mask_store_dto.dart';
+import 'package:modu_3_dart_study/2025-04-04/dto/mask_store_response_dto.dart';
 
-import '../dto/mask_store_response_dto.dart';
 import 'mask_store_data_source.dart';
 
 class MaskStoreDataSourceImpl implements MaskStoreDataSource {
@@ -20,45 +20,45 @@ class MaskStoreDataSourceImpl implements MaskStoreDataSource {
     }
   }
 
-  Future<List<MaskStoreDto>> _fetchFromJsonFile() async {
+  Future<http.Response> _fetchData() async {
     final http.Client client = http.Client();
     try {
       final response = await client.get(Uri.parse(_url));
       if (response.statusCode == 200) {
-        final responseDto = MaskStoreResponseDto.fromJson(
-          json.decode(utf8.decode(response.bodyBytes)),
-        );
-        return responseDto.stores;
+        return response;
       } else {
         throw Exception('데이터 로드 실패: ${response.statusCode}');
       }
     } finally {
       client.close();
     }
+  }
+
+  Future<List<MaskStoreDto>> _fetchFromJsonFile() async {
+    final response = await _fetchData();
+    final responseDto = MaskStoreResponseDto.fromJson(
+      json.decode(utf8.decode(response.bodyBytes)),
+    );
+    return responseDto.stores;
   }
 
   Future<List<MaskStoreDto>> _fetchFromApi() async {
-    final http.Client client = http.Client();
-    try {
-      final response = await client.get(Uri.parse(_url));
-      if (response.statusCode == 200) {
-        return _parseJson(response.body);
-      } else {
-        throw Exception('데이터 로드 실패: ${response.statusCode}');
-      }
-    } finally {
-      client.close();
-    }
+    final response = await _fetchData();
+    return _parseJson(response.body);
   }
 
   List<MaskStoreDto> _parseJson(String responseBody) {
-    final dynamic jsonResponse = json.decode(responseBody);
-    if (jsonResponse is List) {
-      return jsonResponse.map((json) => MaskStoreDto.fromJson(json)).toList();
-    } else if (jsonResponse is Map) {
-      return [MaskStoreDto.fromJson(jsonResponse.cast<String, dynamic>())];
-    } else {
-      throw Exception('Unexpected JSON format');
+    try {
+      final dynamic jsonResponse = json.decode(responseBody);
+      if (jsonResponse is List) {
+        return jsonResponse.map((json) => MaskStoreDto.fromJson(json)).toList();
+      } else if (jsonResponse is Map) {
+        return [MaskStoreDto.fromJson(jsonResponse.cast<String, dynamic>())];
+      } else {
+        throw Exception('Unexpected JSON format');
+      }
+    } catch (e) {
+      throw Exception('JSON 파싱 오류: $e');
     }
   }
 }
