+++ lib/2025-04-04/data_source/mock_data_source/mock_mask_photo_data_source.dart
@@ -1,17 +1,16 @@
-// lib/2025-04-04/data_source/mock_photo_data_source_impl.dart
 import 'dart:convert';
 
-import '../../dto/photo_dto.dart';
-import '../photo_data_source.dart';
+import 'package:modu_3_dart_study/2025-04-04/data_source/photo_data_source.dart';
+import 'package:modu_3_dart_study/2025-04-04/dto/photo_dto.dart';
 
-class MockErrPhotoDataSourceImpl implements PhotoDataSource {
+class MockPhotoDataSourceImpl implements PhotoDataSource {
   @override
   Future<List<PhotoDto>> getPhotos() async {
     // Mock JSON response
     final response = '''
-[
+    [
   {
-    "id": "1",
+    "id": 1,
     "type": "article",
     "title": "This is an article",
     "content": "This is the content of the article.",
@@ -19,13 +18,14 @@ class MockErrPhotoDataSourceImpl implements PhotoDataSource {
   },
   {
     "id": 2,
-    "type": null,
+    "type": "image",
     "url": "https://example.com/image.jpg",
     "caption": "This is an image.",
     "created_at": "2020-02-02"
   },
   {
     "id": 3,
+    "type": "video",
     "url": "https://example.com/video.mp4",
     "caption": "This is a video.",
     "created_at": "2020-03-03"
